dist/datachannel_test_2peer/send/009.html

<!DOCTYPE html>
<html lang="en" class="a">
	<head>
		<meta charset="utf-8">
		<!-- Testharness Timeout to long -> 60 Seconds-->
		<meta name="timeout" content="long">
		<title>WebRTC - Send Test</title>
		<script src="../resources/testharness.js"></script>
		<script src="../resources/testharnessreport.js"></script><link href="../resources/testharness.css" type="text/css" rel="stylesheet"/>
		<!-- Including the Firebase.js to send the ICE Candidate to the other peer-->
		<script type="text/javascript" src='https://cdn.firebase.com/v0/firebase.js'></script>
	</head>
	<body class="a">
		<div>
			<div id="signallingInfo"></div>
			
				
			
		</div>
		<hr />

		<!-- Testharness log -->
		<div id="log"></div>
		<script src="../datachannel2peers.js" type="text/javascript"></script>
		<script type="text/javascript">
            var label = "test";
            var dataChannel = null;
            var expected = "arraybuffer";
			var datasize = 8;
	        var data = new ArrayBuffer(datasize);
	        var int8View = new Int8Array(data);
	        for (var i = 0; i < 8; i++) {
	            int8View[i] = i;
	        }
            var test = async_test("Set up a DataChannel - send an ArrayBufferView (Int8Array)  and receive the data - check type and data");
            if (offerer) {
                test.step(function() {
                    dataChannel = peerConnection.createDataChannel(label);
                    dataChannel.onopen = test.step_func(function() {
                    	try{
                    		dataChannel.send(int8View);
                    	}catch(e){
                    		assert_unreached(e.name + ": " + e.message);
                    	}
                        test.done();
                    });
                });
            } else {
                peerConnection.ondatachannel = test.step_func(function(e) {
                    dataChannel = e.channel;
                    dataChannel.binaryType = expected;
                    dataChannel.onmessage = test.step_func(function(e) {
						var result = new Int8Array(e.data);
            			assert_array_equals(result, int8View, "Array not Equal: ");
			            assert_equals(dataChannel.binaryType, expected, "Send and receive was successfull but Wrong binaryType: ");
			            test.done();
                    });
                });
            }
		</script>
	</body>
</html>

